requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chromadb==0.4.18
pypdf==3.17.4
python-dotenv==1.0.0
streamlit==1.29.0
